+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - [ ] Indication of a rules that are already defined by parent .editorconfig
 - [ ] Indication of rules not matching any files
 - [ ] Validate syntax of sections
-- [x] Light bulb action to sort properties
+- [x] Light bulb to sort properties in a section
+- [x] Light bulb to sort properties in all sections
 - [x] Dismiss QuickInfo on completion sesstion start
 - [x] Context menu command to open settings page
 - [x] New color scheme